requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
pika==1.3.0
psycopg2-binary==2.9.3
pydantic==1.10.2
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
